- added files for the correct GoogleAppEngine Data-class definitions. Some properties are overridden or just declared to something else. Usage now: There are two package.jdo files: The one ending with _NOT-USED is for usage without AppEngine, just rename the package.jdo file to something like this: package.jdo_GAE_NOT-USED, and the other to package.jdo.
- Modified ClassMeta and FieldMeta to have an additional fielIdString for usage with GoogleAppEngine. 
- Modified some other classes to avoid errors on GoogleAppEngine
- Implemented a clone() method to the class FieldMeta, also added a cloning Constructir to it.

+++ org.cumulus4j.store/src/main/java/org/cumulus4j/store/model/EncryptionCoordinateSet.java
@@ -61,7 +61,7 @@ public class EncryptionCoordinateSet
 
 	@PrimaryKey
 	@Persistent(valueStrategy=IdGeneratorStrategy.NATIVE, sequence="EncryptionCoordinateSetSequence")
-	private int encryptionCoordinateSetID = -1;
+	private long encryptionCoordinateSetID = -1;
 
 	@Persistent(nullValue=NullValue.EXCEPTION)
 	private String cipherTransformation;
@@ -109,7 +109,7 @@ public class EncryptionCoordinateSet
 	 * @return the unique numeric identifier (primary key).
 	 */
 	public int getEncryptionCoordinateSetID() {
-		return encryptionCoordinateSetID;
+		return (int)encryptionCoordinateSetID;
 	}
 
 	/**
